% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mclust.R
\name{e_step}
\alias{e_step}
\title{Perform the E-step of the EM algorithm for the GMM}
\usage{
e_step(data, model)
}
\arguments{
\item{data}{The input step for the E-step}

\item{model}{GMM model with initial parameters}
}
\value{
matrix of responsibilities indicating the probablity of each data point belonging to each cluster
}
\description{
Perform the E-step of the EM algorithm for the GMM
}
