% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mclust.R
\name{initialize_gmm}
\alias{initialize_gmm}
\title{Fitting a GMM to a data using the EM algorithm}
\usage{
initialize_gmm(data, K)
}
\arguments{
\item{data}{The input data for GMM initialization}

\item{K}{The number of clusters in the GMM}
}
\value{
a list containing the initial parameter for each cluster in the GMM
}
\description{
Fitting a GMM to a data using the EM algorithm
}
