% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mclust.R
\name{m_step}
\alias{m_step}
\title{perform the M-step of the EM algorithm for the GMM}
\usage{
m_step(data, responsibilities)
}
\arguments{
\item{data}{input data for the M-step}

\item{responsibilities}{matrix obtained from the E-step}
}
\value{
list containing the updated parameter for each cluster in the GMM
}
\description{
perform the M-step of the EM algorithm for the GMM
}
